+++ src/client/view/world_view_settings.cpp
@@ -1,7 +1,7 @@
 #include "client/view/world_view_settings.h"
 
 WorldViewSettings::WorldViewSettings(const size_t& screen_width, const size_t& screen_height,
-                                     float32 world_width, size_t world_height,
+                                     float world_width, float world_height,
                                      const TextureCreator& texture_creator) :
     screen_width(screen_width), screen_height(screen_height),
     world_size(world_width, world_height),
